+++ lib/src/cli_client.dart
@@ -1,6 +1,6 @@
 library dart_coveralls.cli_client;
 
-import 'dart:async' show Future, Completer;
+import 'dart:async' show Future;
 import 'dart:convert';
 import 'dart:io';
 
@@ -15,27 +15,22 @@ import 'process_system.dart';
 import 'services/travis.dart' as travis;
 
 class CommandLineClient {
-  final String projectDirectory;
+  final String packagesPath;
   final String packageRoot;
   final String token;
 
-  CommandLineClient._(this.projectDirectory, this.packageRoot, this.token);
+  CommandLineClient._(this.packagesPath, this.packageRoot, this.token);
 
-  factory CommandLineClient({String projectDirectory, String packageRoot,
+  factory CommandLineClient({String packageRoot, String packagesPath,
       String token, Map<String, String> environment}) {
-    if (projectDirectory == null) {
-      projectDirectory = p.current;
-    }
 
-    packageRoot = _calcPackageRoot(projectDirectory, packageRoot);
-
-    return new CommandLineClient._(projectDirectory, packageRoot, token);
+    return new CommandLineClient._(packagesPath, packageRoot, token);
   }
 
-  Future<CoverageResult<String>> getLcovResult(String testFile,
+  Future<String> getLcovResult(String testFile,
       {int workers, ProcessSystem processSystem: const ProcessSystem()}) {
     var collector =
-        new LcovCollector(packageRoot, processSystem: processSystem);
+        new LcovCollector(packageRoot: packageRoot, packagesPath: packagesPath, processSystem: processSystem);
     return collector.getLcovInformation(testFile, workers: workers);
   }
 
@@ -61,7 +56,10 @@ class CommandLineClient {
     var rawLcov = await getLcovResult(testFile,
         workers: workers, processSystem: processSystem);
 
-    rawLcov.printSummary();
+    if(rawLcov == null){
+      print("Nothing to collect: Connection to VM service timed out. Make sure your test file is free from errors: ${testFile}");
+      exit(0);
+    }
 
     return uploadToCoveralls(rawLcov,
         workers: workers,
@@ -81,7 +79,7 @@ class CommandLineClient {
       bool throwOnConnectivityError: false, int retry: 0,
       bool excludeTestFiles: false, bool printJson}) async {
     var collector =
-        new LcovCollector(packageRoot, processSystem: processSystem);
+        new LcovCollector(packageRoot: packageRoot, packagesPath: packagesPath, processSystem: processSystem);
 
     var result = await collector.convertVmReportsToLcov(containsVmReports,
         workers: workers);
@@ -96,17 +94,20 @@ class CommandLineClient {
         printJson: printJson);
   }
 
-  Future<CoverallsResult> uploadToCoveralls(CoverageResult coverageResult,
+  Future<CoverallsResult> uploadToCoveralls(String coverageResult,
       {int workers, ProcessSystem processSystem: const ProcessSystem(),
       String coverallsAddress, bool dryRun: false,
       bool throwOnConnectivityError: false, int retry: 0,
       bool excludeTestFiles: false, bool printJson}) async {
-    var lcov = LcovDocument.parse(coverageResult.result.toString());
+
+
+    var lcov = LcovDocument.parse(coverageResult);
+
 
     var serviceName = travis.getServiceName(Platform.environment);
     var serviceJobId = travis.getServiceJobId(Platform.environment);
 
-    var report = CoverallsReport.parse(token, lcov, projectDirectory,
+    var report = CoverallsReport.parse(token, lcov, p.current,
         excludeTestFiles: excludeTestFiles,
         serviceName: serviceName,
         serviceJobId: serviceJobId);
@@ -115,7 +116,10 @@ class CommandLineClient {
       print(const JsonEncoder.withIndent('  ').convert(report));
     }
 
-    if (dryRun) return null;
+    if (dryRun) {
+      print("Dry run completed successfully.");
+      return null;
+    }
 
     var endpoint = new CoverallsEndpoint(coverallsAddress);
 
@@ -127,22 +131,9 @@ class CommandLineClient {
       stderr.writeln('Error sending results');
       stderr.writeln(e);
       stderr.writeln(new Chain.forTrace(stack).terse);
+      return null;
     }
   }
-}
-
-String _calcPackageRoot(String packageDir, String packageRoot) {
-  assert(p.isAbsolute(packageDir));
-
-  if (packageRoot == null) {
-    packageRoot = 'packages';
-  }
-
-  if (p.isRelative(packageRoot)) {
-    packageRoot = p.join(packageDir, packageRoot);
-  }
-
-  return p.normalize(packageRoot);
 }
 
 Future<CoverallsResult> _sendLoop(CoverallsEndpoint endpoint, String covString,