+++ lib/src/process_system.dart
@@ -2,7 +2,7 @@ library coveralls.process_system;
 
 import "dart:async" show Future;
 import "dart:convert" show Encoding;
-import "dart:io" show Process, ProcessResult, SYSTEM_ENCODING;
+import "dart:io" show Process, ProcessResult, ProcessStartMode, SYSTEM_ENCODING;
 
 class ProcessSystem {
   const ProcessSystem();
@@ -33,5 +33,17 @@ class ProcessSystem {
         runInShell: runInShell,
         stdoutEncoding: stdoutEncoding,
         stderrEncoding: stderrEncoding);
+  }
+
+  Future<Process> startProcess(String executable, List<String> arguments,
+      {String workingDirectory, Map<String, String> environment,
+      bool includeParentEnvironment: true, bool runInShell: false,
+      ProcessStartMode mode: ProcessStartMode.NORMAL}) {
+    return Process.start(executable, arguments,
+        workingDirectory: workingDirectory,
+        environment: environment,
+        includeParentEnvironment: includeParentEnvironment,
+        runInShell: runInShell,
+        mode: mode);
   }
 }